Vminitd: Add pause command
Due to us supporting a pod type now, and pid ns sharing
being quite a common thing for pods, lets add a pause container
like command to vminitd to eventually enable pid ns sharing between
containers in our variant of a pod.

This changes vminitd slightly to have pause and init (default)
commands as it seemed simpler than creating a whole new binary
to include in the guest image.
